+++ components/persistence/entity-processor-service-impl/src/main/kotlin/net/corda/entityprocessor/impl/internal/ProcessorService.kt
@@ -13,7 +13,7 @@ import net.corda.data.persistence.FindEntities
 import net.corda.data.persistence.FindWithNamedQuery
 import net.corda.data.persistence.MergeEntities
 import net.corda.data.persistence.PersistEntities
-import net.corda.flow.utils.RequestID
+import net.corda.flow.utils.RequestIDGenerator
 import net.corda.flow.utils.toMap
 import net.corda.libs.virtualnode.datamodel.repository.RequestsIdsRepository
 import net.corda.messaging.api.records.Record
@@ -106,19 +106,33 @@ class ProcessorService {
 
         return when (val entityRequest = request.request) {
             is PersistEntities -> {
-                val originalRequestId = request.flowExternalEventContext.requestId
                 val entities = entityRequest.entities
                     .map { serializationService.deserialize(it.array(), Any::class.java) }
 
+                // originalRequestId is set client side based on a combination of entity parameters and suspend count
+                // in order that we can filter out duplicate requests in case this external command was received as the result
+                // of a retry when a previous attempt already succeeded.
+                val originalRequestId = request.flowExternalEventContext.requestId
+
+                // If there is a nondeterministic argument, for example a timestamp, originalRequestId would be different
+                // for retries and not be flagged as duplicate to filter out. To accommodate this we try to modify the request
+                // id hash component to be the hash of the entity id, if there is one, rather than the request parameters.
+                // Any entity ID must be unique, being bound to the primary key of the db table.
                 val requestId = entities
                     .mapNotNull { entityManagerFactory.persistenceUnitUtil.getIdentifier(it) }
                     .takeIf { it.isNotEmpty() }?.let { entityPks ->
-                        RequestID.replaceHash(
+                        RequestIDGenerator.replaceRequestIdHash(
                             originalRequestId,
                             entityPks,
                             serializationService
                         ) { toBase64(it.sha256Bytes()) }
+                        // Where there are no ids to use as a basis for a replacement hash, we fall back to the originalRequestId set
+                        // by the client.
                     } ?: originalRequestId
+
+                // Note if the user code tried to write the same entity twice (as opposed to a retry scenario) the user code should
+                // get an error returned, it shouldn't be filtered as a duplicate. This will work because the suspend count
+                // is written into the request ID and so the request ID would always be different.
 
                 val entityResponse = withDeduplicationCheck(requestId, entityManager, onDuplication = {
                     EntityResponse(emptyList(), KeyValuePairList(emptyList()), null)
